refactor: interface, routes 분리 및 1차 리팩토링
