+++ src/Main/make_goal_pages/make_ai_goal/list.tsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from "react";
 import "./listStyle.css";
 import log from "../img/log.png";
 import { useLocation, useNavigate } from "react-router-dom";
-import MoveBack from "../../../MoneySplit/MoveBack";
-import LoadingSpinner from "../../../MoneySplit/loadingSpinner";
+import MoveBack from "../../../MoneySplit/utils/MoveBack";
+import LoadingSpinner from "../../../MoneySplit/utils/loadingSpinner";
 
 // 추천 목록의 타입 정의
 interface Recommendation {
