+++ src/MoneySplit/pages/ai_pages/ai_anaLoading.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import "../../style/splitStyle.css"; // CSS 파일 import
-import LoadingSpinner from "../../loadingSpinner";
+import LoadingSpinner from "../../utils/loadingSpinner";
 
 const AI_anaLoading: React.FC = () => {
   return (
