+++ src/MoneySplit/pages/final_process/InputPin.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import "../../style/splitStyle.css";
-import MoveBack from "../../MoveBack";
+import MoveBack from "../../utils/MoveBack";
 
 const InputPin: React.FC = () => {
   const [pin, setPin] = useState<string[]>(Array(6).fill(""));
